+++ backend/src/main/java/pl/edu/agh/backend/api/models/CreateShopRequest.java
@@ -1,6 +1,6 @@
-package pl.edu.agh.koluniowezakupki.api.models;
+package pl.edu.agh.backend.api.models;
 
-import pl.edu.agh.koluniowezakupki.db.models.Shop;
+import pl.edu.agh.backend.db.models.Shop;
 
 public record CreateShopRequest(String name, Integer userID) {
     public Shop createShop(){
