+++ backend/src/main/java/pl/edu/agh/backend/api/models/CreateShoppingListRequest.java
@@ -1,6 +1,6 @@
-package pl.edu.agh.koluniowezakupki.api.models;
+package pl.edu.agh.backend.api.models;
 
-import pl.edu.agh.koluniowezakupki.db.models.ShoppingList;
+import pl.edu.agh.backend.db.models.ShoppingList;
 
 public record CreateShoppingListRequest(String name, Integer userID) {
     public ShoppingList createShoppingList(){
