+++ backend/src/main/java/pl/edu/agh/backend/db/models/Group.java
@@ -1,19 +1,17 @@
-package pl.edu.agh.koluniowezakupki.db.models;
+package pl.edu.agh.backend.db.models;
 
 import jakarta.persistence.*;
 import lombok.*;
 
-
 @Entity
-@Table(name = "ListProducts")
-@Builder
+@Table(name = "Groups")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ListProduct {
+public class Group {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer ID;
-    private Integer listID;
-    private Integer productID;
+    private String name;
+    private Integer creatorID;
 }
