+++ backend/src/main/java/pl/edu/agh/backend/db/models/Product.java
@@ -1,22 +1,20 @@
-package pl.edu.agh.koluniowezakupki.db.models;
+package pl.edu.agh.backend.db.models;
 
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.lang.Nullable;
 
 @Entity
-@Table(name = "ShoppingLists")
+@Table(name = "Products")
 @Builder
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ShoppingList {
+public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer ID;
     private String name;
     @Nullable
     private Integer userID;
-
-
 }
