+++ backend/src/main/java/pl/edu/agh/backend/db/models/ShoppingList.java
@@ -1,20 +1,22 @@
-package pl.edu.agh.koluniowezakupki.db.models;
+package pl.edu.agh.backend.db.models;
 
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.lang.Nullable;
 
 @Entity
-@Table(name = "Products")
+@Table(name = "ShoppingLists")
 @Builder
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Product {
+public class ShoppingList {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer ID;
     private String name;
     @Nullable
     private Integer userID;
+
+
 }
