+++ backend/src/main/java/pl/edu/agh/backend/db/ports/ListProductRepository.java
@@ -1,7 +1,7 @@
-package pl.edu.agh.koluniowezakupki.db.ports;
+package pl.edu.agh.backend.db.ports;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import pl.edu.agh.koluniowezakupki.db.models.ListProduct;
+import pl.edu.agh.backend.db.models.ListProduct;
 
 public interface ListProductRepository extends JpaRepository<ListProduct, Integer> {
 }
