+++ backend/src/main/java/pl/edu/agh/backend/db/ports/ShopRepository.java
@@ -1,7 +1,7 @@
-package pl.edu.agh.koluniowezakupki.db.ports;
+package pl.edu.agh.backend.db.ports;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import pl.edu.agh.koluniowezakupki.db.models.Shop;
+import pl.edu.agh.backend.db.models.Shop;
 
 public interface ShopRepository extends JpaRepository<Shop, Integer>{
