+++ backend/src/main/java/pl/edu/agh/backend/db/ports/UserRepository.java
@@ -1,7 +1,7 @@
-package pl.edu.agh.koluniowezakupki.db.ports;
+package pl.edu.agh.backend.db.ports;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import pl.edu.agh.koluniowezakupki.db.models.User;
+import pl.edu.agh.backend.db.models.User;
 
 
 public interface UserRepository extends JpaRepository<User, Integer> {
